pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ytpl-sync"
dynamic = ["version"]
description = "Synchronizes YouTube playlists and channels to local storage and GDrive"
authors = [{name = "Author"}]
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.1.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov"
]

[project.scripts]
ytpl-sync = "ytpl_sync.main:cli"

[tool.setuptools.dynamic]
version = {file = "VERSION"}
